Added event listeners for 'untagged' event
Example has how you'd trigger the 'untagged' event. This
was necessary due to when using draggable's helper:clone setting
the tags no longer are available for dragging out. So this alternative
allows the user to just trigger the event.

Also removed the alt modification as it doesn't have anything to do with
this example